requirements.txt
asgiref==3.8.1
autopep8==2.3.2
black==25.1.0
boto3==1.37.23
botocore==1.37.23
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
cryptography==45.0.4
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-storages==1.14.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gevent==24.11.1
google-auth==2.40.3
greenlet==3.1.1
gunicorn==22.0.0
idna==3.10
jmespath==1.0.1
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9.1
s3transfer==0.11.4
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.6.0
zope.event==5.0
zope.interface==7.2
